新增GUI模块，支持启动常规和全屏模式的图形界面，同时集成Web服务绑定选项，优化跨平台全屏实现逻辑，完善本地服务器连接配置及默认界面设计。

+++ plugins/plug_protocols/p_base/internal/bytes_common_f.go
@@ -37,7 +37,7 @@ func ValidateValueRange(value any, min, max int64) error {
 	if uintVal, ok := value.(uint64); ok {
 		// 检查uint64值是否在int64范围内
 		if uintVal > math.MaxInt64 {
-			return errors.Errorf("uint64 value %d exceeds int64 maximum %d", uintVal, math.MaxInt64)
+			return errors.Errorf("uint64 value %d exceeds int64 maximum %d", uintVal, int64(math.MaxInt64))
 		}
 		val := int64(uintVal)
 		// 修复：正确处理负数范围验证
